Adding new CSV reporter
